// src/types/WorkoutEntry.ts
export interface WorkoutEntry {
  id: string;
  date: string; 
  user: 'Не выбрано' | 'Егор' | 'Сергей' | 'Роман' | 'Слава' | 'Руслан';
  exercise: 'Не выбрано' |'Жим лежа' | 'Жим на наклонной скамье' | 'Штанга на бицепс' | 'Французкий жим' | 'Приседания' | 'Становая тяга' | 'Средняя дельта' | 'Бабочка(грудь)'| 'Подтягивания на турнике';
  weight: number;
  sets: 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 
  reps: 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12 | 13 | 14 | 15 | 16 | 17 | 18 | 19 | 20 ;
  duration:number;
}